requirements.txt
google-generativeai
python-dotenv
langchain
Flask
pypdf
faiss-cpu
langchain_google_genai
langchain_community
requests
Werkzeug